feat: added department router

+++ api/src/controller/employee.controller.ts
@@ -4,11 +4,15 @@ import EmployeeService from "../service/employee.service";
 import { NextFunction, Request, Response, Router } from "express";
 import {
   CreateEmployeeDto,
+  EmployeeResponseDto,
   LoginEmployeeDto,
   UpdateEmployeeDto,
 } from "../dto/employee.dto";
 import { validate } from "class-validator";
 import { authMiddleware } from "../middlewares/auth.middleware";
+import RequestWithUser from "../utils/requestWithUser";
+import { Role } from "../utils/role.enum";
+import { errorFormatter } from "../utils/errorFormatter.utils";
 
 export default class EmployeeController {
   public router: Router;
@@ -16,16 +20,16 @@ export default class EmployeeController {
   constructor(private employeeService: EmployeeService) {
     this.router = Router();
 
-    this.router.get("/", this.getAllEmployees);
+    this.router.get("/", authMiddleware, this.getAllEmployees);
     this.router.get("/:id", authMiddleware, this.getEmployeeById);
     this.router.post("/", authMiddleware, this.createEmployee);
-    this.router.patch("/", this.updateEmployee);
-    this.router.delete("/:id", this.deleteEmployee);
+    this.router.patch("/", authMiddleware, this.updateEmployee);
+    this.router.delete("/:id", authMiddleware, this.deleteEmployee);
     this.router.post("/login", this.loginEmployee);
   }
   getAllEmployees = async (_, res: Response) => {
     const employees = await this.employeeService.getAllEmployees();
-    res.json(employees);
+    res.json(plainToInstance(EmployeeResponseDto, employees));
   };
 
   getEmployeeById = async (req: Request, res: Response, next: NextFunction) => {
@@ -41,28 +45,21 @@ export default class EmployeeController {
     }
   };
 
-  createEmployee = async (req: Request, res: Response, next: NextFunction) => {
+  createEmployee = async (
+    req: RequestWithUser,
+    res: Response,
+    next: NextFunction
+  ) => {
     try {
+      if (req.role !== Role.HR) {
+        throw new HttpException(403, "Invalid Access");
+      }
       const employeeDto = plainToInstance(CreateEmployeeDto, req.body);
       const errors = await validate(employeeDto);
 
       if (errors.length) {
-        let subConstraints = [];
-        let constraints = [];
-
-        if (!errors[0].children.length)
-          constraints = errors.map(
-            ({ constraints }) => Object.values(constraints)[0]
-          );
-        else if (errors[0].children.length)
-          subConstraints = errors[0].children.map(
-            ({ constraints }) => Object.values(constraints)[0]
-          );
-
-        throw new HttpException(400, "Validation Error", [
-          ...constraints,
-          ...subConstraints,
-        ]);
+        const formattedError = errorFormatter(errors);
+        throw new HttpException(400, "Validation Error", formattedError);
       }
 
       const employeeData = await this.employeeService.createNewEmployee(
@@ -73,35 +70,32 @@ export default class EmployeeController {
         employeeDto.password,
         employeeDto.role
       );
-      res.status(201).json(employeeData);
+      res.status(201).json({
+        success: true,
+        message: "Employee Created!",
+        data: plainToInstance(EmployeeResponseDto, employeeData),
+      });
     } catch (error) {
       next(error);
     }
   };
 
-  updateEmployee = async (req: Request, res: Response, next: NextFunction) => {
+  updateEmployee = async (
+    req: RequestWithUser,
+    res: Response,
+    next: NextFunction
+  ) => {
     try {
+      if (req.role !== Role.HR) {
+        throw new HttpException(403, "Invalid Access");
+      }
       const id = Number(req.body.id);
       const employeeDto = plainToInstance(UpdateEmployeeDto, req.body);
       const errors = await validate(employeeDto);
 
       if (errors.length) {
-        let subConstraints = [];
-        let constraints = [];
-
-        if (!errors[0].children.length)
-          constraints = errors.map(
-            ({ constraints }) => Object.values(constraints)[0]
-          );
-        else if (errors[0].children.length)
-          subConstraints = errors[0].children.map(
-            ({ constraints }) => Object.values(constraints)[0]
-          );
-
-        throw new HttpException(400, "Validation Error", [
-          ...constraints,
-          ...subConstraints,
-        ]);
+        const formattedError = errorFormatter(errors);
+        throw new HttpException(400, "Validation Error", formattedError);
       }
 
       await this.employeeService.updateEmployee(id, { ...req.body });
@@ -111,8 +105,15 @@ export default class EmployeeController {
     }
   };
 
-  deleteEmployee = async (req: Request, res: Response, next: NextFunction) => {
+  deleteEmployee = async (
+    req: RequestWithUser,
+    res: Response,
+    next: NextFunction
+  ) => {
     try {
+      if (req.role !== Role.HR) {
+        throw new HttpException(403, "Invalid Access");
+      }
       const id = Number(req.params.id);
       await this.employeeService.deleteEmployee(id);
       res.json({ sucess: true, message: "Employee Deleted!" });
